feat: move and export initialization helper methods
