+++ go/batch-submitter/dial/common.go
@@ -1,0 +1,9 @@
+package dial
+
+import "time"
+
+const (
+	// defaultDialTimeout is default duration the service will wait on
+	// startup to make a connection to either the L1 or L2 backends.
+	defaultDialTimeout = 5 * time.Second
+)
